Fix non-terminating loop in DuplicateContainers::exists

+++ src/vlog/forward/joinprocessor.cpp
@@ -2092,7 +2092,7 @@ bool DuplicateContainers::exists(const Term_t *v) {
         }
     } else {
         if (nfields == 1) {
-            uint8_t idxTable = 0;
+            size_t idxTable = 0;
             size_t emptyTables = 0;
             while (idxTable < ntables) {
                 if (tables[idxTable] != NULL) {
@@ -2115,7 +2115,7 @@ bool DuplicateContainers::exists(const Term_t *v) {
             }
             empty = emptyTables == ntables;
         } else {
-            uint8_t idxTable = 0;
+            size_t idxTable = 0;
             size_t emptyTables = 0;
             while (idxTable < ntables) {
                 if (tables[idxTable] != NULL) {
